feat(server): define `commitments` server and router
