+++ apps/server/package.json
@@ -32,6 +32,7 @@
     "rxjs": "^7.8.1",
     "ts-deepmerge": "^7.0.1",
     "tslib": "^2.7.0",
+    "type-fest": "^4.26.1",
     "zod": "^3.23.8"
   },
   "devDependencies": {
