+++ apps/server/src/app.module.ts
@@ -2,13 +2,12 @@ import { Module } from '@nestjs/common'
 import { AlchemyProxyController } from 'server/alchemy-proxy/alchemy-proxy.controller'
 import { AppController } from 'server/app.controller'
 import { AppService } from 'server/app.service'
-import { SemaphoresService } from 'server/semaphores/semaphores.service'
 import { SupabaseProvider } from 'server/supabase/supabase.provider'
 import { TrpcModule } from 'server/trpc/trpc.module'
 
 @Module({
   imports: [TrpcModule],
   controllers: [AlchemyProxyController, AppController],
-  providers: [AppService, SemaphoresService, SupabaseProvider],
+  providers: [AppService, SupabaseProvider],
 })
 export class AppModule {}
