+++ apps/server/src/commitments/entities/commitment-insert.type.ts
@@ -1,0 +1,3 @@
+import type { Database } from 'server/supabase/supabase.types'
+
+export type CommitmentInsert = Database['public']['Tables']['commitments']['Insert']
