+++ apps/server/src/trpc/trpc.module.ts
@@ -1,7 +1,8 @@
 import { Module } from '@nestjs/common'
 import { BandadaRouter } from 'server/bandada/bandada.router'
 import { BandadaService } from 'server/bandada/bandada.service'
-import { SemaphoresService } from 'server/semaphores/semaphores.service'
+import { CommitmentsRouter } from 'server/commitments/commitments.router'
+import { CommitmentsService } from 'server/commitments/commitments.service'
 import { SupabaseProvider } from 'server/supabase/supabase.provider'
 import { TrpcRouter } from 'server/trpc/trpc.router'
 import { TrpcService } from 'server/trpc/trpc.service'
@@ -10,7 +11,8 @@ import { TrpcService } from 'server/trpc/trpc.service'
   providers: [
     BandadaService,
     BandadaRouter,
-    SemaphoresService,
+    CommitmentsRouter,
+    CommitmentsService,
     SupabaseProvider,
     TrpcService,
     TrpcRouter,
