+++ apps/server/src/trpc/trpc.router.ts
@@ -1,18 +1,21 @@
 import { type INestApplication, Injectable } from '@nestjs/common'
 import * as trpcExpress from '@trpc/server/adapters/express'
 import { BandadaRouter } from 'server/bandada/bandada.router'
+import { CommitmentsRouter } from 'server/commitments/commitments.router'
 import { createContext } from 'server/trpc/trpc.context'
 import { TrpcService } from 'server/trpc/trpc.service'
 
 @Injectable()
 export class TrpcRouter {
   constructor(
+    private readonly bandada: BandadaRouter,
+    private readonly commitments: CommitmentsRouter,
     private readonly trpc: TrpcService,
-    private readonly bandadaRouter: BandadaRouter,
   ) {}
 
   router = this.trpc.router({
-    bandada: this.bandadaRouter.router,
+    bandada: this.bandada.router,
+    commitments: this.commitments.router,
   })
 
   async applyMiddleware(app: INestApplication) {
