+++ supabase/package.json
@@ -1,12 +1,12 @@
 {
   "name": "supabase",
-  "devDependencies": {
-    "supabase": "^1.204.3"
-  },
   "scripts": {
     "restart": "supabase stop && supabase start",
     "start": "supabase start",
     "stop": "supabase stop",
-    "types": "supabase gen types --lang=typescript --local > ../apps/server/src/supabase/supabase.types.ts"
+    "types": "supabase gen types --lang=typescript --local > ../apps/server/src/supabase/supabase-generated.types.ts"
+  },
+  "devDependencies": {
+    "supabase": "^1.204.3"
   }
 }
